refactor: jupyter-server 堅牢化実装とコード改善
Phase 23 の5つの構造的改善を実装:
- 23.1: カーネル単位の実行直列化ロック（asyncio.Lock）
- 23.2: async ハンドラ内の同期 I/O を run_in_executor でオフロード
- 23.3: handlers.py（1515行）を kernel/cell/contents/preview の4モジュールに分割
- 23.4: SQL コネクションプール化（pool_pre_ping + スレッドセーフ化）
- 23.5: sandbox に os.rename/replace/link/symlink のブロック追加

リファクタリング:
- sandbox のパス正規化を realpath に統一（symlink バイパス防止）
- preview_handlers/workspace_handlers の I3 違反（同期 exists()）を解消
- _get_engine に threading.Lock による double-checked locking を追加
- cell_handlers のデッド validate_path ラッパー削除と import 統一
- workspace_handlers の import asyncio をモジュール冒頭に統一
- _sandbox_rename/_sandbox_replace をファクトリ関数で共通化
- _create_sql_engine デッドコードを削除
- KernelExecuteHandler の一般例外を write_error_response に統一
- kernel_handlers の datetime.utcnow() を utc_now_iso() に統一
- test_lock_wait_timeout のテスト実効性を修正
