+++ jupyter-server/extensions/custom_api/contents_handlers.py
@@ -1,0 +1,249 @@
+"""
+ファイル・ノートブック管理 REST API ハンドラー
+"""
+
+from __future__ import annotations
+
+import logging
+from pathlib import Path
+
+from tornado import web
+
+from .base import (
+    JUPYTER_ROOT_DIR,
+    BaseCustomHandler,
+)
+
+log = logging.getLogger(__name__)
+
+
+async def _find_available_path(contents_manager, target_path: str) -> str:
+    """
+    既存ファイルとの重複を避けるため、自動連番でパスを探索する。
+
+    target_path が既に存在する場合:
+      name.ipynb → name_2.ipynb → name_3.ipynb → ...
+    """
+    if not await contents_manager.file_exists(target_path):
+        return target_path
+
+    # ベース名と拡張子に分割
+    filename = target_path.rsplit("/", 1)[-1] if "/" in target_path else target_path
+    if "." in filename:
+        base, ext = target_path.rsplit(".", 1)
+        ext = "." + ext
+    else:
+        base = target_path
+        ext = ""
+
+    counter = 2
+    while counter <= 100:
+        new_path = f"{base}_{counter}{ext}"
+        if not await contents_manager.file_exists(new_path):
+            return new_path
+        counter += 1
+
+    raise ValueError(f"Could not find available path for: {target_path}")
+
+
+async def _create_content(contents_manager, target_path: str, content_type: str = "notebook"):
+    """ノートブックまたはファイルを作成する共通ヘルパー"""
+    if content_type == "notebook":
+        return await contents_manager.new(
+            path=target_path,
+            model={
+                "type": "notebook",
+                "content": {
+                    "cells": [],
+                    "metadata": {},
+                    "nbformat": 4,
+                    "nbformat_minor": 5,
+                },
+            },
+        )
+    return await contents_manager.new(path=target_path)
+
+
+def validate_path(user_input: str, base_dir: str = JUPYTER_ROOT_DIR) -> str:
+    """
+    パストラバーサル攻撃を防ぐためのパス検証
+
+    Args:
+        user_input: ユーザーからの入力パス
+        base_dir: ベースディレクトリ
+
+    Returns:
+        検証済みの相対パス
+
+    Raises:
+        ValueError: 不正なパスの場合
+    """
+    if not user_input:
+        return ""
+
+    # 先頭の / を削除（相対パスとして扱う）
+    clean_path = user_input.lstrip("/")
+
+    # 絶対パス化して検証
+    base = Path(base_dir).resolve()
+    target = (base / clean_path).resolve()
+
+    # ベースディレクトリ配下にあることを確認
+    try:
+        target.relative_to(base)
+    except ValueError:
+        raise ValueError(f"不正なパスです: {user_input}") from None
+
+    return clean_path
+
+
+class ContentsListHandler(BaseCustomHandler):
+    """GET/POST /api/custom/contents"""
+
+    @web.authenticated
+    async def get(self):
+        """ファイル一覧を取得"""
+        path = self.get_argument("path", "/")
+
+        try:
+            # パストラバーサル対策
+            path = validate_path(path)
+            model = await self.contents_manager.get(path, content=True)
+            contents = []
+            if model["type"] == "directory":
+                for item in model.get("content", []):
+                    contents.append(
+                        {
+                            "name": item["name"],
+                            "type": item["type"],
+                            "size": item.get("size"),
+                            "modified_at": item.get("last_modified"),
+                        }
+                    )
+            self.write_success(
+                {
+                    "path": "/" + path if path else "/",
+                    "contents": contents,
+                }
+            )
+        except Exception as e:
+            log.error("Failed to list contents: %s", e, exc_info=True)
+            self.write_error_response("INTERNAL_ERROR", "Failed to list contents", 500)
+
+    @web.authenticated
+    async def post(self):
+        """ノートブックまたはファイルを作成"""
+        body = self.get_json_body()
+        content_type = body.get("type", "notebook")
+        target_path = body.get("path", "")
+
+        try:
+            # パストラバーサル対策
+            target_path = validate_path(target_path)
+            # 既存ファイルとの重複を避ける自動連番
+            target_path = await _find_available_path(self.contents_manager, target_path)
+            model = await _create_content(self.contents_manager, target_path, content_type)
+
+            self.write_success(
+                {
+                    "path": "/" + model["path"],
+                    "type": model["type"],
+                    "created_at": model.get("created") or model.get("last_modified"),
+                }
+            )
+        except Exception as e:
+            log.error("Failed to create content: %s", e, exc_info=True)
+            self.write_error_response("INTERNAL_ERROR", "Failed to create content", 500)
+
+
+class ContentsHandler(BaseCustomHandler):
+    """GET/PUT/DELETE /api/custom/contents/{path}"""
+
+    @web.authenticated
+    async def get(self, path: str):
+        """ファイルまたはノートブックの内容を取得"""
+        try:
+            # パストラバーサル対策
+            path = validate_path(path)
+            model = await self.contents_manager.get(path, content=True)
+            if model["type"] == "notebook":
+                self.write_success(
+                    {
+                        "path": "/" + path,
+                        "type": "notebook",
+                        "content": model["content"],
+                        "modified_at": model.get("last_modified"),
+                    }
+                )
+            else:
+                self.write_success(
+                    {
+                        "path": "/" + path,
+                        "type": model["type"],
+                        "content": model.get("content"),
+                        "modified_at": model.get("last_modified"),
+                    }
+                )
+        except FileNotFoundError:
+            self.write_error_response("NOTEBOOK_NOT_FOUND", f"Not found: {path}", 404)
+        except Exception as e:
+            log.error("Failed to get content '%s': %s", path, e, exc_info=True)
+            self.write_error_response("INTERNAL_ERROR", "Failed to get content", 500)
+
+    @web.authenticated
+    async def post(self, path: str = ""):
+        """ノートブックまたはファイルを作成"""
+        body = self.get_json_body()
+        content_type = body.get("type", "notebook")
+        target_path = body.get("path", path)
+
+        try:
+            # パストラバーサル対策
+            target_path = validate_path(target_path)
+            # 既存ファイルとの重複を避ける自動連番
+            target_path = await _find_available_path(self.contents_manager, target_path)
+            model = await _create_content(self.contents_manager, target_path, content_type)
+
+            self.write_success(
+                {
+                    "path": "/" + model["path"],
+                    "type": model["type"],
+                    "created_at": model.get("created") or model.get("last_modified"),
+                }
+            )
+        except Exception as e:
+            log.error("Failed to create content: %s", e, exc_info=True)
+            self.write_error_response("INTERNAL_ERROR", "Failed to create content", 500)
+
+    @web.authenticated
+    async def put(self, path: str):
+        """ファイルまたはノートブックを更新"""
+        body = self.get_json_body()
+        content = body.get("content")
+
+        try:
+            # パストラバーサル対策
+            path = validate_path(path)
+            model = await self.contents_manager.get(path, content=False)
+            model["content"] = content
+            await self.contents_manager.save(model, path)
+            self.write_success({"path": "/" + path, "status": "updated"})
+        except FileNotFoundError:
+            self.write_error_response("NOTEBOOK_NOT_FOUND", f"Not found: {path}", 404)
+        except Exception as e:
+            log.error("Failed to update content '%s': %s", path, e, exc_info=True)
+            self.write_error_response("INTERNAL_ERROR", "Failed to update content", 500)
+
+    @web.authenticated
+    async def delete(self, path: str):
+        """ファイルまたはノートブックを削除"""
+        try:
+            # パストラバーサル対策
+            path = validate_path(path)
+            await self.contents_manager.delete(path)
+            self.write_success({"path": "/" + path, "status": "deleted"})
+        except FileNotFoundError:
+            self.write_error_response("NOTEBOOK_NOT_FOUND", f"Not found: {path}", 404)
+        except Exception as e:
+            log.error("Failed to delete content '%s': %s", path, e, exc_info=True)
+            self.write_error_response("INTERNAL_ERROR", "Failed to delete content", 500)
