+++ jupyter-server/extensions/custom_api/workspace_handlers.py
@@ -4,6 +4,7 @@
 チャット（AI会話）ごとに独立した作業空間（ワークスペース）の作成・一覧を提供する。
 """
 
+import asyncio
 import json
 import logging
 import uuid
@@ -89,6 +90,81 @@ def _read_workspace(workspace_dir: Path) -> dict | None:
         return None
 
 
+def _create_workspace_sync(root: Path, name: str, summary: str, status: str) -> dict:
+    """ワークスペースのディレクトリ作成とメタデータ書き込みを同期的に行う"""
+    workspace_id = _generate_workspace_id()
+    workspace_dir = root / workspace_id
+    workspace_dir.mkdir(parents=True, exist_ok=True)
+    (workspace_dir / "data").mkdir(exist_ok=True)
+    (workspace_dir / "output").mkdir(exist_ok=True)
+
+    created_at = utc_now_iso()
+    metadata = {
+        "workspace_id": workspace_id,
+        "name": name.strip(),
+        "created_at": created_at,
+        "summary": summary,
+        "status": status,
+    }
+
+    metadata_path = workspace_dir / "metadata.json"
+    with open(metadata_path, "w", encoding="utf-8") as f:
+        json.dump(metadata, f, ensure_ascii=False)
+
+    return {
+        "workspace_id": workspace_id,
+        "name": metadata["name"],
+        "created_at": created_at,
+        "summary": summary,
+        "status": status,
+    }
+
+
+def _list_workspaces_sync(root: Path) -> list[dict]:
+    """ワークスペース一覧を同期的に取得する"""
+    workspaces = []
+    for entry in root.iterdir():
+        if not entry.is_dir():
+            continue
+        ws = _read_workspace(entry)
+        if ws is not None:
+            workspaces.append(ws)
+    workspaces.sort(key=lambda w: w["created_at"], reverse=True)
+    return workspaces
+
+
+def _update_workspace_sync(metadata_path: Path, summary, status) -> dict:
+    """ワークスペースメタデータを同期的に更新する
+
+    Raises:
+        FileNotFoundError: メタデータファイルが存在しない場合
+    """
+    if not metadata_path.exists():
+        raise FileNotFoundError(f"Workspace metadata not found: {metadata_path}")
+
+    with open(metadata_path, encoding="utf-8") as f:
+        metadata = json.load(f)
+
+    if summary is not None:
+        metadata["summary"] = summary
+    if status is not None:
+        metadata["status"] = status
+
+    with open(metadata_path, "w", encoding="utf-8") as f:
+        json.dump(metadata, f, ensure_ascii=False)
+
+    return metadata
+
+
+def _read_templates_sync(templates_dir: Path) -> tuple[str, str]:
+    """テンプレートファイルを同期的に読み込む"""
+    template_path = templates_dir / "summary_template.md"
+    criteria_path = templates_dir / "verification_criteria.md"
+    template = template_path.read_text(encoding="utf-8")
+    verification_criteria = criteria_path.read_text(encoding="utf-8")
+    return template, verification_criteria
+
+
 class WorkspacesHandler(BaseCustomHandler):
     """POST /api/workspaces, GET /api/workspaces"""
 
@@ -117,26 +193,18 @@ class WorkspacesHandler(BaseCustomHandler):
 
         try:
             root = _ensure_workspace_root()
-            workspace_id = _generate_workspace_id()
-            workspace_dir = root / workspace_id
-            workspace_dir.mkdir(parents=True, exist_ok=True)
-            (workspace_dir / "data").mkdir(exist_ok=True)
-            (workspace_dir / "output").mkdir(exist_ok=True)
+            loop = asyncio.get_running_loop()
+            result = await loop.run_in_executor(None, _create_workspace_sync, root, name, summary, status)
 
-            created_at = utc_now_iso()
-            metadata = {
-                "workspace_id": workspace_id,
-                "name": name.strip(),
-                "created_at": created_at,
-                "summary": summary,
-                "status": status,
-            }
-
-            metadata_path = workspace_dir / "metadata.json"
-            with open(metadata_path, "w", encoding="utf-8") as f:
-                json.dump(metadata, f, ensure_ascii=False)
-
-            self.write_success(_format_workspace_info(workspace_id, metadata["name"], created_at, summary, status))
+            self.write_success(
+                _format_workspace_info(
+                    result["workspace_id"],
+                    result["name"],
+                    result["created_at"],
+                    result["summary"],
+                    result["status"],
+                )
+            )
         except Exception as e:
             log.error("Failed to create workspace: %s", e, exc_info=True)
             self.write_error_response("INTERNAL_ERROR", "Failed to create workspace", 500)
@@ -146,17 +214,8 @@ class WorkspacesHandler(BaseCustomHandler):
         """ワークスペース一覧を取得する"""
         try:
             root = _ensure_workspace_root()
-            workspaces = []
-
-            for entry in root.iterdir():
-                if not entry.is_dir():
-                    continue
-                ws = _read_workspace(entry)
-                if ws is not None:
-                    workspaces.append(ws)
-
-            # created_at 降順でソート
-            workspaces.sort(key=lambda w: w["created_at"], reverse=True)
+            loop = asyncio.get_running_loop()
+            workspaces = await loop.run_in_executor(None, _list_workspaces_sync, root)
 
             self.write_success({"workspaces": workspaces})
         except Exception as e:
@@ -196,23 +255,9 @@ class WorkspaceHandler(BaseCustomHandler):
             workspace_dir = root / workspace_id
 
             metadata_path = workspace_dir / "metadata.json"
-            if not metadata_path.exists():
-                self.write_error_response("NOT_FOUND", f"Workspace not found: {workspace_id}", 404)
-                return
 
-            # 既存メタデータを読み込み
-            with open(metadata_path, encoding="utf-8") as f:
-                metadata = json.load(f)
-
-            # 指定されたフィールドのみ更新
-            if summary is not None:
-                metadata["summary"] = summary
-            if status is not None:
-                metadata["status"] = status
-
-            # 書き戻し
-            with open(metadata_path, "w", encoding="utf-8") as f:
-                json.dump(metadata, f, ensure_ascii=False)
+            loop = asyncio.get_running_loop()
+            metadata = await loop.run_in_executor(None, _update_workspace_sync, metadata_path, summary, status)
 
             self.write_success(
                 _format_workspace_info(
@@ -223,6 +268,8 @@ class WorkspaceHandler(BaseCustomHandler):
                     metadata.get("status", "not_started"),
                 )
             )
+        except FileNotFoundError:
+            self.write_error_response("NOT_FOUND", f"Workspace not found: {workspace_id}", 404)
         except json.JSONDecodeError:
             log.error("Corrupted metadata.json for workspace %s", workspace_id)
             self.write_error_response("INTERNAL_ERROR", "Workspace metadata is corrupted", 500)
@@ -243,24 +290,26 @@ class WorkspaceSummarizeHandler(BaseCustomHandler):
             self.write_error_response("VALIDATION_ERROR", ws_error, 400)
             return
 
-        # ワークスペースの存在確認
+        # ワークスペースの存在確認 + テンプレートファイル読み込み（スレッドプールで実行）
         root = _ensure_workspace_root()
         workspace_dir = root / workspace_id
         metadata_path = workspace_dir / "metadata.json"
-        if not metadata_path.exists():
-            self.write_error_response("NOT_FOUND", f"Workspace not found: {workspace_id}", 404)
-            return
 
-        # テンプレートファイル読み込み
+        def _check_workspace_and_read_templates():
+            if not metadata_path.exists():
+                raise FileNotFoundError(f"Workspace not found: {workspace_id}")
+            return _read_templates_sync(_TEMPLATES_DIR)
+
         try:
-            template_path = _TEMPLATES_DIR / "summary_template.md"
-            criteria_path = _TEMPLATES_DIR / "verification_criteria.md"
-
-            template = template_path.read_text(encoding="utf-8")
-            verification_criteria = criteria_path.read_text(encoding="utf-8")
+            loop = asyncio.get_running_loop()
+            template, verification_criteria = await loop.run_in_executor(None, _check_workspace_and_read_templates)
         except FileNotFoundError as e:
-            log.error("Template file not found: %s", e)
-            self.write_error_response("INTERNAL_ERROR", "Template file not found", 500)
+            error_msg = str(e)
+            if "Workspace not found" in error_msg:
+                self.write_error_response("NOT_FOUND", error_msg, 404)
+            else:
+                log.error("Template file not found: %s", e)
+                self.write_error_response("INTERNAL_ERROR", "Template file not found", 500)
             return
         except Exception as e:
             log.error("Failed to read template files: %s", e, exc_info=True)
